Fix up TX avail calculation.
We now use the number of BDs used from the original BD bd_cnt field.
There is no reason to look at the idx now... though maybe in the
future, we'll need to use it in case we get aggregated packets out
of TXBD order.

The previous commit was incorrect.