Fix invalid TX BDs, handle errors better
Long TSO BDs were getting random garbage in their flags_type value due to the
length hint lookup table not being long enough.  Stop using the table for sizes
over 2048 bytes.

Check the completion type of TX completions, and log errors

If we are using an offload (including TSO), we will use two TX BDs for the
first frag and one for each after that.  Since we can use a max of 32 BDs, this
makes our effective segment limit 31, not 32.

Use the second TX ring for the completion queue for now since the first one
appears to be the only one getting bus DMA sync calls.